Require less tedious input from user by automatically creating a new game if no saves exist, or displaying saves if they do exist. Fix save numbering system.

+++ lib/save.rb
@@ -1,7 +1,9 @@
 require 'yaml'
 
 module Save
-  @@saves = Dir.entries("output").length - 2
+  # Get save number so that they don't repeat
+  # Rescue value of 0 in case no saved games exist (can't read nil array)
+  @@saves = Dir.entries("output").select {|name| /yaml/.match(name)}.sort[-1][5].to_i rescue 0
 
   def save_game
     Dir.mkdir("output") unless Dir.exists?("output")
@@ -12,15 +14,26 @@ module Save
   end
 
   def load_game
-    files = display_saves
+    files = get_files
     if !files.empty?
-      puts "Enter the number of the game you'd like to resume:"
-      file_number = gets.chomp.to_i - 1
-      save = files[file_number]
-      from_yaml(save)
-      File.delete("output/#{save}") if File.exist?("output/#{save}")
+      display_saves
+      puts "Enter the number of the game you'd like to resume, or enter 'new' for a new game:"
+      input = gets.chomp
+      if input == "new"
+        new_game
+      elsif input 
+        begin
+          file_number = input.to_i - 1
+          save = files[file_number]
+          puts "Opening #{save}..."
+          from_yaml(save)
+          File.delete("output/#{save}") if File.exist?("output/#{save}")
+        rescue
+        puts "Game not found. Starting a new game..."
+        new_game
+        end
+      end
     else
-      puts "No saves found. Starting a new game..."
       new_game
     end
   end
@@ -29,13 +42,15 @@ module Save
     name = "save_#{@@saves + 1}"
   end
 
+  def get_files
+    Dir.entries("output").select {|name| /yaml/.match(name)}.sort
+  end
+
   def display_saves
-    files = Dir.entries("output").select {|name| /yaml/.match(name)}.sort
     puts "Game #\tFile Name(s)"
-    files.each_with_index do |name, index|
+    get_files.each_with_index do |name, index|
       puts "#{index + 1}\t#{name.to_s}"
     end
-    files
   end
 
   def to_yaml
